refactor: Type restructuring
Move types to more agnostic placements
