+++ src/core/types.ts
@@ -1,5 +1,6 @@
 import type { StandardSchemaV1 } from '@standard-schema/spec';
-import type { Options, SSEClientSourceInit } from '../types';
+import type { Options } from '../types/request';
+import type { SSEClientSourceInit } from '../types/sse';
 import type { SafeWrap } from '../utils/wrap';
 
 // biome-ignore lint/suspicious/noExplicitAny: This is used for inferrence, and requires any so inference works as it should
